ClaraBackend/requirements.txt
Django==2.2.1
django-cors-headers==2.5.3
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
mysqlclient==1.4.2.post1
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
